Corrections to Weather_Cost.xlsx
The temperature from KNMI weather stations is typically given in 0.1°C, so the values in the Excel file appeared to be off by a factor of 10.

Also, the column RH appeared to contain different data in WeatherData.xlsx than Weather_Cost.xlsx. I have corrected the values in Weather_Cost.xlsx to match the form of WeatherData.xlsx